pyproject.toml
[project]
name = "ai-trends-explorer"
version = "0.1.0"
description = "Multi-agent system that collects AI research papers, news, and signals — produces weekly briefings and monthly deep reports."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3",
    "langgraph>=0.4",
    "langchain-anthropic>=0.3",
    "langchain-groq>=0.3",
    "langchain-google-genai>=2.1",
    "langsmith>=0.3",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    "sentence-transformers>=3.3",
    "arxiv>=2.1",
    "feedparser>=6.0",
    "httpx>=0.28",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "resend>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_service"]

[tool.ruff]
target-version = "py312"
line-length = 100
